+++ docs/changelog.md
@@ -2,6 +2,11 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+* Add the ability to add images to operator messages like a dialog box.
+* Add non-blocking mode for operator message.
+* Add `clear_operator_message` function for closing operator message.
+* Change socket data processing.
+
 ## HardPy 0.8.0
 
 * Modify API for dialog boxes with images.